feat(radio): enhance genres and artists seeding functionality; update slug field requirements and refactor migration files
