+++ src/modules/Content/Radio/seed/seedArtists/artists.ts
@@ -1,0 +1,90 @@
+export const artists = [
+  {
+    name: 'Alex Stein',
+    slug: 'alex-stein',
+    // photo: './img/alex-stein-2.jpg',
+    photo:
+      'https://res.cloudinary.com/mohmdevcloud/image/upload/v1733303209/aammiiroo/artists/icjcdbarpfp0uwyy5alk.jpg',
+    bio: 'A renowned techno DJ and producer known for his melodic yet hard-hitting tracks.',
+    Tracks: {
+      docs: [1, 5, 7], // IDs for related Radio tracks: "Aurora", "Pulse", "Genesis"
+    },
+  },
+  {
+    name: 'Victor Ruiz',
+    slug: 'victor-ruiz',
+    // photo: './img/victor-ruiz-1.jpg',
+    photo:
+      'https://res.cloudinary.com/mohmdevcloud/image/upload/v1733303208/aammiiroo/artists/zfvmb5puxwhfpubyjfpa.jpg',
+    bio: 'A Brazilian techno artist recognized for his powerful and emotional productions.',
+    Tracks: {
+      docs: [2], // ID for the track: "Neural Collapse"
+    },
+  },
+  {
+    name: 'ARTBAT',
+    slug: 'artbat',
+    // photo: './img/artbat-1.jpg',
+    photo:
+      'https://res.cloudinary.com/mohmdevcloud/image/upload/v1733303205/aammiiroo/artists/smazhs7wzbhfq7y4eyma.jpg',
+    bio: 'Ukrainian duo creating emotive, melodic techno for the global stage.',
+    Tracks: {
+      docs: [3], // ID for the track: "Afterlife"
+    },
+  },
+  {
+    name: 'Charlotte de Witte',
+    slug: 'charlotte-de-witte',
+    // photo: './img/charlotte-de-witte-1.jpg',
+    photo:
+      'https://res.cloudinary.com/mohmdevcloud/image/upload/v1733303205/aammiiroo/artists/ytjk4seprq0kbwf19mvt.jpg',
+    bio: 'A leading figure in the techno scene, known for her hard-hitting tracks and powerful performances.',
+    Tracks: {
+      docs: [4], // ID for the track: "Eclipse"
+    },
+  },
+  {
+    name: 'Enrico Sangiuliano',
+    slug: 'enrico-sangiuliano',
+    // photo: './img/enrico-sangiuliano-1.jpg',
+    photo:
+      'https://res.cloudinary.com/mohmdevcloud/image/upload/v1733303205/aammiiroo/artists/cpbxdrsfix2htqekyzsp.jpg',
+    bio: 'An Italian producer renowned for his expertly crafted and atmospheric techno tracks.',
+    Tracks: {
+      docs: [6], // ID for the track: "Stratosphere"
+    },
+  },
+  {
+    name: 'Amelie Lens',
+    slug: 'amelie-lens',
+    // photo: './img/amelie-lens-2.jpg',
+    photo:
+      'https://res.cloudinary.com/mohmdevcloud/image/upload/v1733303205/aammiiroo/artists/qoj0yipl66kquymm7waw.jpg',
+    bio: 'A Belgian techno artist celebrated for her hypnotic sounds and energetic DJ sets.',
+    Tracks: {
+      docs: [8], // ID for the track: "Reverie"
+    },
+  },
+  {
+    name: 'Tale of Us',
+    slug: 'tale-of-us',
+    // photo: './img/tale-of-us-1.jpg',
+    photo:
+      'https://res.cloudinary.com/mohmdevcloud/image/upload/v1733303205/aammiiroo/artists/lto1nixjxfi5p7lnuee6.jpg',
+    bio: 'An Italian duo known for their cinematic and emotional techno compositions.',
+    Tracks: {
+      docs: [9], // ID for the track: "Dreamstate"
+    },
+  },
+  {
+    name: 'Reinier Zonneveld',
+    slug: 'reinier-zonneveld',
+    // photo: './img/reinier-zonneveld-1.jpg',
+    photo:
+      'https://res.cloudinary.com/mohmdevcloud/image/upload/v1733303205/aammiiroo/artists/zkwkpbvg6qhejvtbgyzj.jpg',
+    bio: 'A Dutch techno producer famous for his energetic live performances and powerful beats.',
+    Tracks: {
+      docs: [10], // ID for the track: "Momentum"
+    },
+  },
+]
